+++ lib/src/core/preference_key.dart
@@ -4,6 +4,7 @@ import 'pref_types.dart';
 class PreferenceKey<T> {
   final String key;
   final PreferenceStorage storage;
+  final String? writePolicy;
   final T? defaultValue;
   final String description;
   final PrefSerializer<T>? serializer;
@@ -11,6 +12,7 @@ class PreferenceKey<T> {
   const PreferenceKey({
     required this.key,
     required this.storage,
+    this.writePolicy,
     this.defaultValue,
     this.description = '',
     this.serializer,